moving some codes to Cookie container
