+++ src/containers/Cookie/Cookie.js
@@ -1,18 +1,124 @@
 import React, { Component } from 'react';
-// import { getQuotes } from '../../cleaners/getQuotes';
-// import leftCookie from '../../images/left-cookie.png';
-// import rightCookie from '../../images/right-cookie.png';
-// import { keepInJar } from '../../actions';
-// import { withRouter } from 'react-router-dom';
-// import { connect } from 'react-redux';
-// import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import { keepWisdom } from '../../actions';
+import leftCookie from '../../images/left-cookie.png';
+import rightCookie from '../../images/right-cookie.png';
+import wisdom from '../../images/wisdom-paper.jpg';
+import toJar from '../../images/put-inna-jar.png';
+import PropTypes from 'prop-types';
+import './Cookie.css';
 
 export class Cookie extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      revealWisdom: false,
+      brokenCookie: false, 
+      wisdom: {},
+      jar: []
+    }
+  }
+
+  eatCookie = () => {
+    const { wisdoms } = this.props;
+
+    this.setState({ 
+      revealWisdom: !this.state.revealWisdom,
+      brokenCookie: !this.state.brokenCookie
+     })
+    this.state.revealWisdom === true ? this.hideWisdom(wisdoms) : this.showWisdom(wisdoms);
+  }; 
+
+  showWisdom = async (wisdoms) => {
+    const randomKey = Math.floor(Math.random() * wisdoms.length);
+    const wisdom = wisdoms[randomKey];
+    this.setState({ wisdom });
+  };
+
+  hideWisdom = () => {
+    this.setState({ 
+      wisdom: {},
+      brokenCookie: false,
+      revealWisdom: false });
+  };
+
+  putWisdomInJar = () => {
+    let { jar } = this.props;
+    let { wisdom } = this.state;
+    
+    if (!jar.find(wisdomInJar => wisdomInJar.id === wisdom.id)) {
+      this.props.keepWisdom(wisdom);
+    }
+    this.setState({ 
+      revealWisdom: false,
+      brokenCookie: false 
+    });
+  }
 
   render() {
+    let leftCookieState = this.state.brokenCookie === true ? 'left-cookie-broken' : '';
+    let rightCookieState = this.state.brokenCookie === true ? 'right-cookie-broken' : '';
+
     return (
-     <div></div> 
+     <div>
+          
+      { !this.state.broken && 
+        <div className='cookie'>
+          <img src={leftCookie}
+          className={`left-cookie ${leftCookieState}`}
+          alt="left piece of cookie"
+          onClick={this.eatCookie} />  
+          <img src={rightCookie}
+          className={`right-cookie ${rightCookieState}`}
+          alt="right piece of cookie"
+           onClick={this.eatCookie} /> 
+        </div>
+      }
+
+      { this.state.revealWisdom &&
+        <div>
+          <div className='wisdom-wrapper'>
+            <p className='message'>
+              {this.state.wisdom.message}
+            </p>
+            <img src={wisdom}
+              className='wisdom-paper'
+              alt="Wisdom on a paper"
+              onClick={this.eatCookie} />
+          </div>   
+          <div className='putjar-btn'>
+          <img
+            src={toJar}
+            className='to-jar'
+            alt="A jar to throw in wisdoms"
+            onClick={this.putWisdomInJar} />
+          <p className='btn-txt'>Keep wisdom in a jar</p>
+          </div>
+        </div>      
+      } 
+
+    </div> 
     )
   }
 }
  
+export const mapStateToProps = (state) => {
+  return {
+    wisdoms: state.wisdoms,
+    jar: state.jar
+  };
+};
+
+export const mapDispatchToProps = (dispatch) => {
+  return {
+    keepWisdom: (wisdom) => (dispatch(keepWisdom(wisdom)))
+  };
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(Cookie);
+
+Cookie.propTypes = {
+  keepWisdom: PropTypes.func,
+  wisdoms: PropTypes.array,
+  jar: PropTypes.array
+};
